#30 Autorización
Se establecen Políticas y roles

+++ app/Http/Controllers/PetitionController.php
@@ -14,14 +14,17 @@ class PetitionController extends Controller
     ];
 
     public function index(){
+        $this->authorize('viewAny', Petition::class);
         return new PetitionCollection(Petition::paginate(10));
     }
 
     public function show(Petition $petition){
+        $this->authorize('view', $petition);
         return response()->json(new PetitionResource($petition), 200);
     }
 
     public function store(Request $request){
+        $this->authorize('create', Petition::class);
         $validatedData = $request->validate([
             'status' => 'required',
         ], self::$messages);
@@ -30,11 +33,13 @@ class PetitionController extends Controller
     }
 
     public function update(Request $request, Petition $petition){
+        $this->authorize('update', $petition);
         $petition->update($request->all());
         return response()->json($petition, 200);
     }
 
     public function delete(Request $request, Petition $petition){
+        $this->authorize('delete', $petition);
         $petition->delete();
         return response()->json(null, 204);
     }
